+++ packages/conc/lib/Polysemy/Conc/Gate.hs
@@ -1,0 +1,10 @@
+{-# options_haddock prune #-}
+
+-- |Description: API for the 'Gate' effect
+module Polysemy.Conc.Gate (
+  module Polysemy.Conc.Effect.Gate,
+  module Polysemy.Conc.Interpreter.Gate,
+) where
+
+import Polysemy.Conc.Effect.Gate
+import Polysemy.Conc.Interpreter.Gate