+++ packages/conc/lib/Polysemy/Conc/Sync.hs
@@ -43,6 +43,7 @@ whileEmptyInterval interval action =
       whenM (not <$> Sync.empty @a) (Time.sleep @t @d interval *> spin)
 
 -- |Run an action with a locally scoped 'Sync' variable.
+--
 -- This avoids a dependency on @'Embed' 'IO'@ in application logic while still allowing the variable to be scoped.
 withSync ::
   ∀ d res r .