+++ packages/conc/changelog.md
@@ -7,6 +7,7 @@
 * Add `timeoutStop`, a variant of `timeout` that calls `Polysemy.Resume.stop`.
 * Add `Lock`, a mutex effect.
 * Add `withAsyncTrigger`, a variant of `withAsync` that requires the async action to signal the sync action to start.
+* Add `Gate`, an synchronization point effect.
 
 # 0.9.0.0